backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-multipart==0.0.9

# Security
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# GCP and Gemini
google-cloud-storage>=2.14.0
google-generativeai>=0.4.0

# Celery and Redis
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
pypdf==4.0.1
email-validator==2.1.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Neo4j Graph Database
neo4j==5.16.0

# LangGraph and LangChain
langgraph==0.0.26
langchain-core>=0.1.25,<0.2.0

# Utilities
python-dotenv==1.0.0
